[rust] Add lk-sys
The lk-sys crate provides low-level bindings to LK functions and types.
This runs bindgen in build.rs, and includes the generated output in the
main library. In addition, there are a small number of instances that
need to be in this file, as these must be in the same crate as the
definitions.

Signed-off-by: David Brown <david.brown@linaro.org>