[rust] Build core/alloc
Instead of depending on a pre-defined target tuple, with a pre-build
core and allow, capture our configuration with a custom LLVM target JSON
file. To do this, we need to build core and alloc, which currently
requires a nightly compiler.

The advantage here is that core and alloc will be build with the exact
configuration we define here, including any specific command line
arguments.

Signed-off-by: David Brown <david.brown@linaro.org>