[rust][arm64] Build properly for softfloat
Base this json file off of aarch64-unknown-none-softfloat, to properly
disable use of HW floating point registers in the kernel.

This avoids needing to explicitly pass and option to avoid using Neon.

Note that at the time of this commit, building core with this
configuration generates a warning, due to some explicitly neon code in
core.  Presumably, this will be fixed in core at least before the
compiler starts considering this to be an error.

Signed-off-by: David Brown <davidb@davidb.org>