[rust] Clean up HAVE/USE RUST configuration
Cleanly separate enablement of Rust support from a `USE_RUST` that can
be used when invoking make, and `HAVE_RUST` that is defined internally
when rust support is enabled. This allows targets to decide for
themselves if Rust support is needed, or if it is optional.

The definitive decision about rust support will be determined if a
platform makefile includes `lib/rust_support/rules.mk` to include Rust
support in the build. It can optional do this based on a use-defined
`USE_RUST` flag.

Signed-off-by: David Brown <davidb@davidb.org>