[rust] Allow modules to declare themselves as Rust crates
The variable RUST_CRATES contains a list of paths (from the build root)
to crates to add. For each crate: generate a call to
`cratename::must_link()` to force the crate to be linked in, allowing lk
init blocks to be included.

This makes the pl011 crate a module, so the crate will be included when
the module is included.

Signed-off-by: David Brown <davidb@davidb.org>