[lib][uefi] Add include/uefi to include path list
When we import the GBL UEFI headers from upstream, we put them under a
"include/uefi" subdirectory. However the headers still include each
other with relative paths without the subdir. As a result relative
includes (#include with "") are broken, and we needed to patch the
imported header files.

Previously I added symlink alias to work around the include path issue.
But I think this is not very sustainable and feels hacky.
Instead, just add the include/uefi subdir to the include path list.