[build] allow modules with no source files defined
Allows the overall logic of the build system to only include the module
once, and allow other parts of the build system such as adding include
directories. Otherwise none of the rest of the module logic is used,
such as the generation of the local config.mk or adding any .o file to
the final system linkage.