[ci][rust] Test the x86_64 build as well
Add this target as well for Rust support.

We can remove the toolchain install entirely.  The image already has
rustup installed, and the wrapped cargo it provides will read the
toolchain file and install the correct version of the rust tools.  This
keeps us from having to track down the exact version of Rust needed.

Change-Id: I64d635d97bfaf5fdf72189228a60176adff1d678